Changed address type to uint32_t for clarity

+++ src/revolveLib/Address.cpp
@@ -14,7 +14,7 @@ Address getBaseAddress()
 		throw std::runtime_error("Could not get base address");
 	}
 
-	return static_cast<Address>(reinterpret_cast<uintptr_t>(hModule));
+	return static_cast<Address>(reinterpret_cast<uint32_t>(hModule));
 }
 
 Address getDynamicAddress(Address relativeAddress)
@@ -24,6 +24,6 @@ Address getDynamicAddress(Address relativeAddress)
 
 Address operator+(Address lhs, Address rhs)
 {
-	return static_cast<Address>(static_cast<unsigned int>(lhs) + static_cast<unsigned int>(rhs));
+	return static_cast<Address>(static_cast<uint32_t>(lhs) + static_cast<std::uint32_t>(rhs));
 }
 } // namespace revolve
